- Add two new config options to StandardTable.
### defaultCellRenderer
This can be used to override default text renderer for the whole table.
Column cellRenderer will still override this. Since this is global for the table, and not local to a column,
the type of itemValue can not be known. It should use `label` as source of text instead.
If omitted, TextCell is used as usual.

### defaultTextSize
This config specifies the text size to be used by the default TextCell component.
If omitted, "normal" is used.