- Click outside to close modal now only reacts to onMouseDown.
# Use case

User presses mouse button inside modal, then moves pointer outside of modal and releases the mouse button.
The click is registered as outside the modal and closes it. Not very nice!

Now it uses onMouseDown instead of onClick, which requires the mouse down event to come from outside of modal.